% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_order.R
\name{ANNt_order}
\alias{ANNt_order}
\title{classify assets by the probability of return exceeding a RM
Use artificial neural networks and t-distribution, the number of input neurons is the number of import assets}
\usage{
ANNt_order(
  Initial_Date_Training,
  Final_Date_Training,
  Final_Date_Testing,
  Hidden,
  Stepmax
)
}
\arguments{
\item{Final_Date_Training}{Series end Date (If '' is the System Date)}

\item{Hidden}{Number of hidden neurons (If '' is the length series)}

\item{Stepmax}{Number of replications per asset to train the ANN}

\item{Inicial_Date}{Series start Date (Must be 7 perios greater than the analyzed seriess)}
}
\description{
classify assets by the probability of return exceeding a RM
Use artificial neural networks and t-distribution, the number of input neurons is the number of import assets
}
\examples{
Initial_Date_Training <-c('2018-01-11')
Final_Date_Training <- c('2022-12-29')
Final_Date_Testing <-c('2023-09-07')
Hidden <- 5
Stepmax <- 2500
ANNt_order ('2018-01-11','2022-12-29','', 5, 2500)
# Estimated processing time 30 minutes per asset

}
