% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assets_series.R
\name{Assets_series}
\alias{Assets_series}
\title{ANNt
Generate portfolios with classification by the probability of return exceeding a RM,
using artificial neural networks and t-distribution.}
\usage{
Assets_series(Tickers, RM, Initial_Date, Final_Date)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets}

\item{RM}{Proxy of the market}

\item{Final_Date}{Series end Date ('Year-Month-Day')
Assets with values not observedin the series are excluded}

\item{Inicial_Date}{Series start Date, format ('Year-Month-Day')}
}
\description{
ANNt
Generate portfolios with classification by the probability of return exceeding a RM,
using artificial neural networks and t-distribution.
}
\examples{
# Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
Tickers <-c('AAPL','GOOG','CCBG', 'XOM', 'TSLA')
RM <-c('^GSPC') #RM the S&P500
Initial_Date <-c('2018-01-03')
Final_Date <-c('2023-09-07')

# Generates the Adjusted Daily Prices Series from Yahoo Finance
Assets_series (Tickers=c('AAPL','GOOG','CCBG','XOM','TSLA'),'^GSPC', '2018-01-03', '2023-01-09')

}
