% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_portfolios.R
\name{Gen_portfolios}
\alias{Gen_portfolios}
\title{Generate return portfolios}
\usage{
Gen_portfolios(N_assets, Initial_Date_Testing, Final_Date_Testing, Rf)
}
\arguments{
\item{N_assets}{Number of assets at the ANNt and MF-DFA portfolios}

\item{Initial_Date_Testing}{Initial date of the portfolios series}

\item{Final_Date_Testing}{Final date of the portfolios series}

\item{Rf}{Return of Risk free % (daily)}
}
\description{
Generate return portfolios
}
\examples{
N_assets <- 3
Initial_Date_Testing <- c('2023-01-03')
Final_Date_Testing <- c('2023-09-07')
Rf <- 0

# Generate the portfolios ANNt,MF-DFA, Sharpe, and Marowitz
Gen_portfolios(3, '2023-01-03', '',0)

}
