% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_tickers.R
\name{Gen_tickers}
\alias{Gen_tickers}
\title{Generate tickers}
\usage{
Gen_tickers(Generate_Tickers)
}
\description{
Generate tickers
}
\examples{
# Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
# Apply the "load" command and use in "Asset_series" command
Generate_Tickers <-c('PETR4.SA','AAPL','Current_SP500_Tickers')

# Generates the Adjusted Daily Prices Series from Yahoo Finance
Gen_tickers (Generate_Tickers=c('PETR4.SA','AAPL','Current_SP500_Tickers'))
load('Generate_Tickers.rda')
}
