% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_backtesting.R
\name{Portfolio_backtesting}
\alias{Portfolio_backtesting}
\title{Portfolio_backtesting
Realize the portfolio backtesting}
\usage{
Portfolio_backtesting()
}
\arguments{
\item{()}{No require parameters}
}
\description{
Portfolio_backtesting
Realize the portfolio backtesting
}
\examples{
# example code
Portfolio_backtesting()

}
