% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assets_series.R
\name{Assets_series}
\alias{Assets_series}
\title{Import the series}
\usage{
Assets_series(Tickers, RM, Initial_Date, Final_Date, Periodicity)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets}

\item{RM}{Proxy of the market}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day')

Assets with values not observed in the series are excluded}

\item{Final_Date}{Series end Date ('Year-Month-Day')}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”, “hourly”, “1minutes”, “2minutes”, “5minutes”, “15minutes”, “30minutes”, “60minutes”, “90minutes”. (Intraday maximum 7 days)}
}
\description{
Import the series
}
\examples{
# Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
Tickers <-c('AAPL','GOOG','CCBG', 'XOM', 'TSLA')
RM <-c('^GSPC') #RM the S&P500
Initial_Date <-c('2018-01-03')
Final_Date <-c('2023-09-07')
Periodicity <- c('daily')

# Generates the Adjusted Daily Prices Series from Yahoo Finance
Assets_series (Tickers=c('AAPL','GOOG','CCBG','XOM','TSLA'),'^GSPC', '2018-01-03', '2023-09-07','daily')

}
