% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_portfolios.R
\name{Gen_portfolios}
\alias{Gen_portfolios}
\title{Gen_portfolios}
\usage{
Gen_portfolios(N_Assets, Initial_Date_Testing, Final_Date_Testing, Rf)
}
\arguments{
\item{N_Assets}{limit of asset numbers in the portfolio}

\item{Initial_Date_Testing}{Initial Date of Test Series}

\item{Final_Date_Testing}{Final Date Test Series}

\item{Rf}{Risk free rate}
}
\description{
Gen_portfolios
}
\examples{
N_Assets <- 3
Initial_Date_Testing <- c('2023-01-03')
Final_Date_Testing <- c('2023-09-07')
Rf <- 0

# Generate assets portfolio (maximum N assets specified)
Gen_portfolios(3,'2023-01-03','',0)

}
